pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wotd"
version = "0.1.0"
description = "AI Word Of The Day — daily trending AI term from RSS, newsletters, and X."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "djalmaaraujo" }]
dependencies = [
    "feedparser>=6.0.10",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "readability-lxml>=0.8.1",
    "lxml>=5.0",
    "PyYAML>=6.0",
    "jinja2>=3.1",
    "python-dateutil>=2.9",
    "anthropic>=0.39",
    "duckdb>=1.1",
    "pyarrow>=17.0",
    "polars>=1.0",
    "jsonschema>=4.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
]

[project.scripts]
wotd = "wotd.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wotd = ["resources/*.txt", "sources/*.txt"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
